refactor(sync): split read_plugin_assets out of claude adapter
The 2,045-line `claude.rs` was dominated by one trait method body:
`read_plugin_assets` walked the plugin cache layout
`cache/<marketplace>/<plugin>/<version>/` and was 205 LOC of
filtering and asset construction — the largest single function in
the workspace's largest file.

Two changes here:
1. Convert `claude.rs` into a `claude/` directory module so future
   topical extracts have a home (skills, commands, settings,
   instructions in subsequent PRs).
2. Move `read_plugin_assets` into `claude/plugin_assets.rs` as a
   free function. The trait method becomes a one-line delegation.
   The helper accesses the adapter via a new
   `pub(in crate::adapters::claude) fn config_root_ref(&self)`
   accessor (PathBuf borrow, no clone) and a `pub(super)`
   `semver_tuple` already in `mod.rs`.

claude/mod.rs:    1853 LOC (was 2045)
claude/plugin_assets.rs: 237 LOC (new, includes module docstring)

316 sync-lib tests pass.
